requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
curl-cffi==0.5.9
duckduckgo-search==3.9.6
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
frozenlist==1.4.0
g4f==0.1.8.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
mmh3==4.0.1
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
Pillow==10.1.0
platformdirs==4.0.0
py-arkose-generator==0.0.0.2
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
PyExecJS==1.5.1
pyjsparser==2.7.1
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
starlette==0.27.0
tiktoken==0.5.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
waitress==2.1.2
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
